Move all rpacket_tracker tests to one file
