+++ tardis/transport/montecarlo/tests/test_rpacket_tracker.py
@@ -9,6 +9,31 @@ from tardis.transport.montecarlo.r_packet import InteractionType
 from tardis.base import run_tardis
 
 
+@pytest.fixture(scope="module")
+def config_rpacket_tracker(example_configuration_dir):
+    """Config object for rpacket tracker"""
+    return Configuration.from_yaml(
+        example_configuration_dir / "tardis_configv1_verysimple.yml"
+    )
+
+
+@pytest.fixture(scope="module")
+def simulation_rpacket_tracking_enabled(config_rpacket_tracker, atomic_dataset):
+    """Simulation object with track_rpacket_enabled"""
+    config_rpacket_tracker.montecarlo.iterations = 1
+    config_rpacket_tracker.montecarlo.no_of_packets = 4000
+    config_rpacket_tracker.montecarlo.last_no_of_packets = -1
+    config_rpacket_tracker.montecarlo.tracking.track_rpacket = True
+    config_rpacket_tracker.spectrum.num = 2000
+    atomic_data = deepcopy(atomic_dataset)
+    sim = run_tardis(
+        config_rpacket_tracker,
+        atom_data=atomic_data,
+        show_convergence_plots=False,
+    )
+    return sim
+
+
 @pytest.fixture()
 def interaction_type_last_interaction_class(
     simulation_rpacket_tracking_enabled,
@@ -117,3 +142,48 @@ def test_rpacket_tracker_properties(expected, obtained, request):
     expected = request.getfixturevalue(expected)
     obtained = request.getfixturevalue(obtained)
     npt.assert_allclose(expected, obtained)
+
+
+def test_rpacket_trackers_to_dataframe(simulation_rpacket_tracking_enabled):
+    sim = simulation_rpacket_tracking_enabled
+    transport_state = sim.transport.transport_state
+    rtracker_df = rpacket_trackers_to_dataframe(transport_state.rpacket_tracker)
+
+    # check df shape and column names
+    assert rtracker_df.shape == (
+        sum([len(tracker.r) for tracker in transport_state.rpacket_tracker]),
+        8,
+    )
+    npt.assert_array_equal(
+        transport_state.rpacket_tracker_df.columns.values,
+        np.array(
+            [
+                "status",
+                "seed",
+                "r",
+                "nu",
+                "mu",
+                "energy",
+                "shell_id",
+                "interaction_type",
+            ]
+        ),
+    )
+
+    # check all data with rpacket_tracker
+    expected_rtrackers = []
+    for rpacket in transport_state.rpacket_tracker:
+        for rpacket_step_no in range(len(rpacket.r)):
+            expected_rtrackers.append(
+                [
+                    rpacket.status[rpacket_step_no],
+                    rpacket.seed,
+                    rpacket.r[rpacket_step_no],
+                    rpacket.nu[rpacket_step_no],
+                    rpacket.mu[rpacket_step_no],
+                    rpacket.energy[rpacket_step_no],
+                    rpacket.shell_id[rpacket_step_no],
+                    rpacket.interaction_type[rpacket_step_no],
+                ]
+            )
+    npt.assert_array_equal(rtracker_df.to_numpy(), np.array(expected_rtrackers))
